requirements.txt
daytona
python-dotenv
fastapi
uvicorn
websockets
aiohttp
pydantic
GitPython
google-generativeai
python-multipart
jinja2

